fix: relocate pack page
