+++ app/packs/@modal/(.)id/[id]/page.tsx
@@ -1,18 +1,18 @@
 'use client';
 import { use } from "react";
 
-import { PackDetails } from '../../_components/QuestionPackDetails'
+import { PackDetails } from '../../../_components/QuestionPackDetails'
 import { Dialog, DialogContent } from '@/components/ui/dialog'
 import { Drawer, DrawerContent } from '@/components/ui/drawer'
 import { useIsMobile } from '@/hooks/use-mobile'
 import { useRouter } from 'next/navigation'
 
 export default function Page(props: { params: Promise<{ id: string }> }) {
-    const params = use(props.params);
-    const router = useRouter()
-    const isMobile = useIsMobile()
+	const params = use(props.params);
+	const router = useRouter()
+	const isMobile = useIsMobile()
 
-    if (isMobile) {
+	if (isMobile) {
 		return (
 			<Drawer
 				defaultOpen
@@ -32,7 +32,7 @@ export default function Page(props: { params: Promise<{ id: string }> }) {
 		)
 	}
 
-    return (
+	return (
 		<Dialog
 			defaultOpen
 			onOpenChange={() => {
