+++ app/packs/id/[id]/page.tsx
@@ -9,7 +9,6 @@ import {
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
 import { ArrowLeft, CopyIcon, FileUp, Heart, LinkIcon } from 'lucide-react'
 import { PackDetailsSkeleton } from './_components/PackDetailsSkeleton'
-import ExportQuestionModal from '../_components/ExportQuestionModal'
 import { useEffect, useState, use } from 'react';
 import type { PackResponse } from '@/types/Packs'
 import { Button } from '@/components/ui/button'
@@ -24,20 +23,21 @@ import axios from 'axios'
 import clsx from 'clsx'
 import { languageMap } from '@/helpers'
 import { packMap } from '@/types'
+import ExportQuestionModal from '../../_components/ExportQuestionModal'
 
 export default function PackDetails(props: { params: Promise<{ id: string }> }) {
-    const params = use(props.params);
-    const router = useRouter()
-    const [packToShow, setPackToShow] = useState<PackResponse | null>(null)
-    const [userData, setUserData] = useState({
+	const params = use(props.params);
+	const router = useRouter()
+	const [packToShow, setPackToShow] = useState<PackResponse | null>(null)
+	const [userData, setUserData] = useState({
 		username: 'Private User',
 		avatar: '/Logo.png',
 		id: undefined
 	})
-    const [searchQuery, setSearchQuery] = useState('')
-    const [isLoading, setIsLoading] = useState(true)
+	const [searchQuery, setSearchQuery] = useState('')
+	const [isLoading, setIsLoading] = useState(true)
 
-    useEffect(() => {
+	useEffect(() => {
 		async function getPack() {
 			setIsLoading(true) // Set loading to true before fetching data
 			const res = await fetch(`/api/packs/${params.id}`)
@@ -63,12 +63,12 @@ export default function PackDetails(props: { params: Promise<{ id: string }> }) 
 		getPack()
 	}, [params.id])
 
-    const filteredQuestions =
+	const filteredQuestions =
 		packToShow?.data.questions.filter((question) =>
 			question.question.toLowerCase().includes(searchQuery.toLowerCase())
 		) ?? []
 
-    const toggleLike = async () => {
+	const toggleLike = async () => {
 		try {
 			const response = await axios.put(`/api/packs/${params.id}/likes`)
 
@@ -82,7 +82,7 @@ export default function PackDetails(props: { params: Promise<{ id: string }> }) 
 						likes:
 							updatedLikes.userLiked ?
 								prev!.data.likes.filter((id) => id !== updatedLikes.id)
-							:	[...prev!.data.likes, updatedLikes.id]
+								: [...prev!.data.likes, updatedLikes.id]
 					},
 					likes: updatedLikes.likes,
 					userLiked: updatedLikes.userLiked
@@ -93,25 +93,25 @@ export default function PackDetails(props: { params: Promise<{ id: string }> }) 
 		}
 	}
 
-    const copyShareLink = () => {
+	const copyShareLink = () => {
 		navigator.clipboard.writeText(
 			`${process.env.NEXT_PUBLIC_PAGE_URL}/packs/${params.id}`
 		)
 		toast.success('Copied to clipboard!')
 	}
 
-    const copyCommand = () => {
+	const copyCommand = () => {
 		navigator.clipboard.writeText(
 			`/import ${packToShow?.data.type} ${params.id}`
 		)
 		toast.success('Copied to clipboard!')
 	}
 
-    return (
+	return (
 		<>
 			{isLoading ?
 				<PackDetailsSkeleton />
-			:	<TooltipProvider delayDuration={0}>
+				: <TooltipProvider delayDuration={0}>
 					<Container>
 						<div className="grid grid-cols-1 lg:grid-cols-4 gap-4">
 							<div className="bg-card rounded-lg p-4 border space-y-4 lg:sticky lg:top-28 h-max @container">
